requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.5
Pillow>=8.2.0
matplotlib>=3.4.2
scikit-learn>=0.24.2
tqdm>=4.61.0
